requirements.txt
langchain
langchain-community
langchain-anthropic
python-dotenv
pypdf
